\name{raw.output.H3K4me3}
\docType{data}
\alias{raw.output.H3K4me3}
\title{Example raw.output object from H3K4me3 data}
\description{
	This data set is an example of the output obtained from the \code{\link{bayespeak}} function. The ChIP-seq experiment in question investigates H3K4me4 methylation data, in a short region of mouse chromosome 16.

	The bed files used to generate this object are included in the BayesPeak package - please see the examples section in the \code{\link{bayespeak}} help file for the code that can be used to recreate this object.

	The raw output has not been truncated in any way.
}
\usage{raw.output.H3K4me3}
\format{A list of 3 objects. See \code{\link{bayespeak}} for more details.}
%%\source{FIXME}
\references{
Many thanks to Dr. Duncan Odom's group for permission to use this data set.
}
\keyword{datasets}

