\name{raw.output}
\docType{data}
\alias{raw.output}
\title{Example raw.output object}
\description{
	This data set is an example of the output obtained from the bayespeak() function, in particular as an example of job parameters. The ChIP-seq experiment in question investigates ER binding in cells from the MCF7 cell line.

	To keep the size of BayesPeak down, raw.output$peaks has been truncated - only the peaks on chromosome 16 are given. raw.output$QC has not been truncated in any way.
}
\usage{raw.output}
\format{A list of 3 objects. See \code{\link{bayespeak}} for more details.}
%%\source{FIXME}
\references{
Many thanks to Dr. Jason Carroll's group for permission to use this data set.
}
\keyword{datasets}

